requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
click==8.1.8
dateparser==1.2.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.3
pydantic_core==2.33.1
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.40
Telethon==1.40.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
Werkzeug==3.1.3
yarl==1.20.0
PyJWT[crypto]==2.10.1
cryptography>=42